+++ lib/Elastica/Aggregation/Filters.php
@@ -59,6 +59,30 @@ class Filters extends AbstractAggregation
     }
 
     /**
+     * @param bool $otherBucket
+     *
+     * @return $this
+     */
+    public function setOtherBucket($otherBucket)
+    {
+        if (!is_bool($otherBucket)) {
+            throw new \InvalidArgumentException('other_bucket only supports boolean values');
+        }
+
+        return $this->setParam('other_bucket', $otherBucket);
+    }
+
+    /**
+     * @param string $otherBucketKey
+     *
+     * @return $this
+     */
+    public function setOtherBucketKey($otherBucketKey)
+    {
+        return $this->setParam('other_bucket_key', $otherBucketKey);
+    }
+
+    /**
      * @return array
      */
     public function toArray()
@@ -73,6 +97,14 @@ class Filters extends AbstractAggregation
             } else {
                 $array['filters']['filters'][] = current($filter)->toArray();
             }
+        }
+
+        if ($this->hasParam('other_bucket')) {
+            $array['filters']['other_bucket'] = $this->getParam('other_bucket');
+        }
+
+        if ($this->hasParam('other_bucket_key')) {
+            $array['filters']['other_bucket_key'] = $this->getParam('other_bucket_key');
         }
 
         if ($this->_aggs) {